requirements.txt
opencv-python
mediapipe
spotipy
python-dotenv 